+++ tests/call.test.js
@@ -1,24 +1,25 @@
 import {
   afterEach, beforeEach, expect, test,
 } from '@jest/globals';
-import { getPrompt, handleEvents, removePrompt } from '../app/index.js';
-import { COMMAND_DEACTIVATE } from '../constants/command.js';
+import {
+  getPrompt, handleEvents, removePrompt, settings,
+} from '../app/index.js';
+import config from '../config/index.js';
 import { SETTING_AI_ACTIVATED } from '../constants/setting.js';
 import storage from '../storage/index.js';
 import { createEvents, TIMEOUT, USER_ID } from './utils.js';
 
 beforeEach(() => {
-  storage.setItem(SETTING_AI_ACTIVATED, true);
+  storage.setItem(SETTING_AI_ACTIVATED, false);
 });
 
 afterEach(() => {
   removePrompt(USER_ID);
 });
 
-test('COMMAND_DEACTIVATE', async () => {
+test('COMMAND_CALL', async () => {
   const events = [
-    ...createEvents([COMMAND_DEACTIVATE.text]),
-    ...createEvents(['嗨']), // should be ignored
+    ...createEvents([`${config.SETTING_AI_NAME} 你好`]),
   ];
   let results;
   try {
@@ -26,11 +27,11 @@ test('COMMAND_DEACTIVATE', async () => {
   } catch (err) {
     console.error(err);
   }
-  expect(getPrompt(USER_ID).lines.length).toEqual(1 * 2);
+  expect(getPrompt(USER_ID).lines.length).toEqual(3 * 2);
   const replies = results.map(({ messages }) => messages.map(({ text }) => text));
   expect(replies).toEqual(
     [
-      [COMMAND_DEACTIVATE.reply],
+      ['OK!'],
     ],
   );
 }, TIMEOUT);