+++ tests/command.test.js
@@ -8,11 +8,10 @@ import {
   getPrompt, handleEvents, removePrompt, settings,
 } from '../app/index.js';
 import { COMMAND_COMMAND } from '../constants/command.js';
-import storage from '../storage/index.js';
 import { createEvents, TIMEOUT, USER_ID } from './utils.js';
 
 beforeEach(() => {
-  storage.initialize(settings);
+  //
 });
 
 afterEach(() => {